moved jax stuff into Jax(Explicit/Implicit)Component
