+++ openmdao/components/implicit_func_comp.py
@@ -135,7 +135,7 @@ class ImplicitFuncComp(ImplicitComponent):
             kwargs = _copy_with_ignore(meta, omf._allowed_add_output_args, ignore=('resid',))
             self.add_output(name, **kwargs)
 
-    def _setup_jax(self, from_group=False):
+    def _setup_jax(self):
         # TODO: this is here to prevent the ImplicitComponent base class from trying to do its
         # own jax setup if derivs_method is 'jax'. We should probably refactor this...
         pass
