+++ openmdao/components/jax_explicit_comp.py
@@ -3,9 +3,14 @@ An ExplicitComponent that uses JAX for derivatives.
 """
 
 import sys
+import inspect
+from types import MethodType
+
 from openmdao.core.explicitcomponent import ExplicitComponent
-from openmdao.utils.jax_utils import jax
 from openmdao.utils.om_warnings import issue_warning
+from openmdao.utils.jax_utils import jax, jit, ExplicitCompJaxify, \
+    compute_partials as _jax_compute_partials, \
+    compute_jacvec_product as _jax_compute_jacvec_product, ReturnChecker, _jax_register_pytree_class
 
 
 class JaxExplicitComponent(ExplicitComponent):
@@ -34,3 +39,75 @@ class JaxExplicitComponent(ExplicitComponent):
             issue_warning(f"{self.msginfo}: JAX is not available, so '{fallback_derivs_method}' "
                           "will be used for derivatives.")
             self.options['derivs_method'] = fallback_derivs_method
+
+    def _setup_jax(self):
+        """
+        Set up the jax interface for this component.
+        """
+        if self.matrix_free is True:
+            self.compute_jacvec_product = MethodType(_jax_compute_jacvec_product, self)
+        else:
+            self.compute_partials = MethodType(_jax_compute_partials, self)
+            self._has_compute_partials = True
+
+        if self.compute_primal is None:
+            # convert the compute method to a compute_primal method
+            jaxifier = ExplicitCompJaxify(self, verbose=True)
+
+            if jaxifier.get_self_statics:
+                self.get_self_statics = MethodType(jaxifier.get_self_statics, self)
+            # replace existing compute method with base class method, so that compute_primal
+            # will be called.
+            self.compute = MethodType(ExplicitComponent.compute, self)
+
+            self.compute_primal = MethodType(jaxifier.compute_primal, self)
+            self._compute_primal_returns_tuple = True
+        else:
+            # check that compute_primal args are in the correct order
+            args = list(inspect.signature(self.compute_primal).parameters)
+            if args and args[0] == 'self':
+                args = args[1:]
+            compargs = self._get_compute_primal_argnames()
+            if args != compargs:
+                raise RuntimeError(f"{self.msginfo}: compute_primal method args {args} don't match "
+                                   f"the expected args {compargs}.")
+
+            # determine if the compute_primal method returns a tuple
+            self._compute_primal_returns_tuple = ReturnChecker(self.compute_primal).returns_tuple()
+
+        if self.options['use_jit']:
+            static_argnums = []
+            idx = len(self._var_rel_names['input']) + 1
+            static_argnums.extend(range(idx, idx + len(self._discrete_inputs)))
+            self.compute_primal = MethodType(jit(self.compute_primal.__func__,
+                                                 static_argnums=static_argnums), self)
+
+        _jax_register_pytree_class(self.__class__)
+
+    def _get_jac_func(self):
+        """
+        Return the jacobian function for this component.
+
+        In forward mode, jax.jacfwd is used, and in reverse mode, jax.jacrev is used.  The direction
+        is chosen automatically based on the sizes of the inputs and outputs.
+
+        Returns
+        -------
+        function
+            The jacobian function.
+        """
+        # TODO: modify this to use relevance and possibly compile multiple jac functions depending
+        # on DV/response so that we don't compute any derivatives that are always zero.
+        if self._jac_func_ is None:
+            fjax = jax.jacfwd if self.best_partial_deriv_direction() == 'fwd' else jax.jacrev
+            nstatic = len(self._discrete_inputs)
+            wrt_idxs = list(range(1, len(self._var_abs2meta['input']) + 1))
+            self._jac_func_ = MethodType(fjax(self.compute_primal.__func__, argnums=wrt_idxs), self)
+
+            if self.options['use_jit']:
+                static_argnums = tuple(range(1 + len(wrt_idxs), 1 + len(wrt_idxs) + nstatic))
+                self._jac_func_ = MethodType(jit(self._jac_func_.__func__,
+                                                 static_argnums=static_argnums),
+                                             self)
+
+        return self._jac_func_
