+++ openmdao/components/jax_implicit_comp.py
@@ -3,9 +3,13 @@ An ImplicitComponent that uses JAX for derivatives.
 """
 
 import sys
+import inspect
+from types import MethodType
+
 from openmdao.core.implicitcomponent import ImplicitComponent
-from openmdao.utils.jax_utils import jax
 from openmdao.utils.om_warnings import issue_warning
+from openmdao.utils.jax_utils import jax, jit, \
+    linearize as _jax_linearize, apply_linear as _jax_apply_linear, _jax_register_pytree_class
 
 
 class JaxImplicitComponent(ImplicitComponent):
@@ -34,3 +38,49 @@ class JaxImplicitComponent(ImplicitComponent):
             issue_warning(f"{self.msginfo}: JAX is not available, so "
                           f"'{fallback_derivs_method}' will be used for derivatives.")
             self.options['derivs_method'] = fallback_derivs_method
+
+    def _setup_jax(self):
+        if self.matrix_free is True:
+            self.apply_linear = MethodType(_jax_apply_linear, self)
+        else:
+            self.linearize = MethodType(_jax_linearize, self)
+            self._has_linearize = True
+
+        if self.compute_primal is None:
+            raise RuntimeError(f"{self.msginfo}: compute_primal is not defined for this component.")
+
+        # check that compute_primal args are in the correct order
+        args = list(inspect.signature(self.compute_primal).parameters)
+        if args and args[0] == 'self':
+            args = args[1:]
+        compargs = self._get_compute_primal_argnames()
+        if args != compargs:
+            raise RuntimeError(f"{self.msginfo}: compute_primal method args {args} don't match "
+                                f"the expected args {compargs}.")
+
+        if self.options['use_jit']:
+            static_argnums = []
+            idx = len(self._var_rel_names['input']) + len(self._var_rel_names['output']) + 1
+            static_argnums.extend(range(idx, idx + len(self._discrete_inputs)))
+            self.compute_primal = MethodType(jit(self.compute_primal.__func__,
+                                                 static_argnums=static_argnums), self)
+
+        _jax_register_pytree_class(self.__class__)
+
+    def _get_jac_func(self):
+        # TODO: modify this to use relevance and possibly compile multiple jac functions depending
+        # on DV/response so that we don't compute any derivatives that are always zero.
+        if self._jac_func_ is None:
+            fjax = jax.jacfwd if self.best_partial_deriv_direction() == 'fwd' else jax.jacrev
+            wrt_idxs = list(range(1, len(self._var_abs2meta['input']) +
+                                  len(self._var_abs2meta['output']) + 1))
+            primal_func = self.compute_primal.__func__
+            self._jac_func_ = MethodType(fjax(primal_func, argnums=wrt_idxs), self)
+
+            if self.options['use_jit']:
+                static_argnums = tuple(range(1 + len(wrt_idxs), 1 + len(wrt_idxs) +
+                                             len(self._discrete_inputs)))
+                self._jac_func_ = MethodType(jit(self._jac_func_.__func__,
+                                                 static_argnums=static_argnums),
+                                             self)
+        return self._jac_func_
