+++ openmdao/jax/tests/test_jax_implicit.py
@@ -68,52 +68,6 @@ class JaxQuadraticCompPrimal(om.JaxImplicitComponent):
         return a * x ** 2 + b * x + c
 
 
-
-class JaxLinearSystemCompConverted(om.ImplicitComponent):
-
-    def initialize(self):
-        self.options.declare('size', default=1, types=int)
-
-    def setup(self):
-        size = self.options['size']
-
-        shape = (size, )
-
-        self.add_input("A", val=np.eye(size))
-        self.add_input("b", val=np.ones(shape))
-        self.add_output("x", shape=shape, val=.1)
-
-    def setup_partials(self):
-        size = self.options['size']
-        mat_size = size * size
-        full_size = size
-
-        row_col = np.arange(full_size, dtype="int")
-
-        self.declare_partials('x', 'b', val=np.full(full_size, -1.0), rows=row_col, cols=row_col)
-
-        rows = np.repeat(np.arange(full_size), size)
-
-        cols = np.arange(mat_size)
-
-        self.declare_partials('x', 'A', rows=rows, cols=cols)
-
-        cols = np.tile(np.arange(size), size)
-        cols += np.repeat(np.arange(1), mat_size) * size
-
-        self.declare_partials(of='x', wrt='x', rows=rows, cols=cols)
-
-        if self.matrix_free:
-            self.linear_solver = om.ScipyKrylov()
-        else:
-            self.linear_solver = om.DirectSolver()
-        self.nonlinear_solver = om.NewtonSolver(solve_subsystems=False)
-
-    def apply_nonlinear(self, inputs, outputs, residuals):
-        residuals['x'] = inputs['A'].dot(outputs['x']) - inputs['b']
-
-
-
 class JaxLinearSystemCompPrimal(om.JaxImplicitComponent):
 
     def initialize(self):
@@ -254,33 +208,6 @@ class JaxLinearSystemCompPrimalwDiscrete(om.JaxImplicitComponent):
 @unittest.skipIf(jax is None or sys.version_info < (3, 9), 'jax is not available or python < 3.9.')
 class TestJaxImplicitComp(unittest.TestCase):
 
-    @parameterized.expand(itertools.product(['fwd', 'rev'], ['matfree', '']), name_func=parameterized_name)
-    def test_quad_comp_converted(self, mode, matrix_free):
-        p = om.Problem()
-        # create an IVC manually so we can set the shapes.  Otherwise must set shape in the component
-        # itself.
-        ivc = p.model.add_subsystem('ivc', om.IndepVarComp('a'))
-        ivc.add_output('b')
-        ivc.add_output('c')
-        comp = p.model.add_subsystem('comp', QuadraticComp(derivs_method='jax'))
-        comp.matrix_free = bool(matrix_free)
-        p.model.connect('ivc.a', 'comp.a')
-        p.model.connect('ivc.b', 'comp.b')
-        p.model.connect('ivc.c', 'comp.c')
-        p.setup(mode=mode)
-
-        p.set_val('ivc.a', 1.0)
-        p.set_val('ivc.b', -4.0)
-        p.set_val('ivc.c', 3.0)
-
-        p.final_setup()
-        p.run_model()
-
-        assert_near_equal(p.get_val('comp.x'), 3.0)
-        assert_check_totals(p.check_totals(of=['comp.x'], wrt=['comp.a', 'comp.b', 'comp.c'], method='fd',
-                                           show_only_incorrect=True, abs_err_tol=3e-5, rel_err_tol=5e-6), atol=3e-5, rtol=5e-6)
-        assert_check_partials(p.check_partials(show_only_incorrect=True))
-
     @parameterized.expand(itertools.product(['fwd', 'rev'], ['matfree', ''], [(), (3,), (2,3)]), name_func=parameterized_name)
     def test_jax_quad_comp(self, mode, matrix_free, shape):
         p = om.Problem()
@@ -307,36 +234,6 @@ class TestJaxImplicitComp(unittest.TestCase):
         assert_check_totals(p.check_totals(of=['comp.x'], wrt=['comp.a', 'comp.b', 'comp.c'], method='fd',
                                            show_only_incorrect=True, abs_err_tol=3.5e-5, rel_err_tol=5e-6), atol=3.5e-5, rtol=5e-6)
         assert_check_partials(p.check_partials(show_only_incorrect=True, abs_err_tol=3e-6, rel_err_tol=3e-6), atol=2.5e-6)
-
-    @parameterized.expand(itertools.product(['fwd', 'rev'], ['matfree', '']), name_func=parameterized_name)
-    def test_lin_system_converted(self, mode, matrix_free):
-        A = np.array([[1., 1., 1.], [1., 2., 3.], [0., 1., 3.]])
-        b = np.array([1, 2, -3])
-
-        prob = om.Problem()
-
-        ivc = prob.model.add_subsystem('ivc', om.IndepVarComp())
-        ivc.add_output('A', A)
-        ivc.add_output('b', b)
-
-        lingrp = prob.model.add_subsystem('lingrp', om.Group())
-        lin = lingrp.add_subsystem('lin', JaxLinearSystemCompConverted(size=3, derivs_method='jax'))
-        lin.matrix_free = bool(matrix_free)
-
-        prob.model.connect('ivc.A', 'lingrp.lin.A')
-        prob.model.connect('ivc.b', 'lingrp.lin.b')
-
-        prob.setup(mode=mode)
-        prob.set_solver_print(level=0)
-
-        prob.run_model()
-
-        assert_near_equal(prob['lingrp.lin.x'], np.array([-4, 9, -4]), tolerance=1e-14)
-
-        assert_check_totals(prob.check_totals(of=['lingrp.lin.x'], wrt=['ivc.b', 'ivc.A'],
-                                              abs_err_tol=2e-4, rel_err_tol=3e-6, show_only_incorrect=True),
-                            atol=2e-4, rtol=3e-6)
-        assert_check_partials(prob.check_partials(show_only_incorrect=True), rtol=1e-5)
 
     @parameterized.expand(itertools.product(['fwd', 'rev'], ['matfree', '']), name_func=parameterized_name)
     def test_jax_lin_system_primal(self, mode, matrix_free):
